Revert "edited server"
