+++ index.html
@@ -15,10 +15,9 @@
                 <h3>Directions</h3>
                 <input type="text" placeholder="D-100" id="from">
                 <input type="text" placeholder="K-05" id="to">
-
                 <button onclick="runScript()">Run Script</button>
                 <div id="output"></div>
-                <script src="js/server.js"></script>
+                <script type="module" src="js/server.js"></script>
             </section>
 
             <section id="about">hi</section>
