--- js/school-map.json
@@ -1,21 +1,0 @@
-{
-    "1": [
-        { "node": 2, "dist": 7 },
-        { "node": 4, "dist": 3 }
-    ],
-    "2": [
-        { "node": 1, "dist": 7 },
-        { "node": 3, "dist": 2 }
-    ],
-    "3": [
-        { "node": 2, "dist": 2 },
-        { "node": 5, "dist": 11 }
-    ],
-    "4": [
-        { "node": 1, "dist": 3 },
-        { "node": 3, "dist": 9 }
-    ],
-    "5": [
-        { "node": 3, "dist": 11 }
-    ]
-}
